refactor(admin): rewire admin shell and routes to feature-scoped modules
AdminApp.vue and the app entry point still imported pages, composables, log state and scss from the old flat layout. Point them at the feature-scoped folders created by the admin reorganisation so the new module structure is wired end to end.
