+++ webui/src/main.ts
@@ -5,29 +5,29 @@ import "tdesign-vue-next/dist/tdesign.css";
 
 import { bootstrapGraphEditor } from "./graph_editor_bootstrap";
 import AdminApp from "./admin/AdminApp.vue";
-import Dashboard from "./admin/view/Dashboard.vue";
-import Connections from "./admin/view/Connections.vue";
-import ConnectionManager from "./admin/view/ConnectionManager.vue";
-import Llm from "./admin/view/Llm.vue";
-import AgentConfig from "./admin/view/AgentConfig.vue";
-import RoleService from "./admin/view/RoleService.vue";
-import Graphs from "./admin/view/Graphs.vue";
-import Tasks from "./admin/view/Tasks.vue";
-import ScheduledTasks from "./admin/view/ScheduledTasks.vue";
-import Logs from "./admin/view/Logs.vue";
-import Commands from "./admin/view/Commands.vue";
-import Chat from "./admin/view/Chat.vue";
-import DataExplorer from "./admin/view/DataExplorer.vue";
-import DataExplorerDetail from "./admin/view/DataExplorerDetail.vue";
-import Settings from "./admin/view/Settings.vue";
-import ApiKeys from "./admin/view/ApiKeys.vue";
-import SetupWizard from "./admin/view/SetupWizard.vue";
-import Plugins from "./admin/view/Plugins.vue";
+import Dashboard from "./admin/dashboard/Dashboard.vue";
+import Connections from "./admin/connections/Connections.vue";
+import ConnectionManager from "./admin/connection-manager/ConnectionManager.vue";
+import Llm from "./admin/llm/Llm.vue";
+import AgentConfig from "./admin/agents/AgentConfig.vue";
+import RoleService from "./admin/agents/RoleService.vue";
+import Graphs from "./admin/graphs/Graphs.vue";
+import Tasks from "./admin/tasks/Tasks.vue";
+import ScheduledTasks from "./admin/scheduled-tasks/ScheduledTasks.vue";
+import Logs from "./admin/logs/Logs.vue";
+import Commands from "./admin/commands/Commands.vue";
+import Chat from "./admin/chat/Chat.vue";
+import DataExplorer from "./admin/data-explorer/DataExplorer.vue";
+import DataExplorerDetail from "./admin/data-explorer/DataExplorerDetail.vue";
+import Settings from "./admin/settings/Settings.vue";
+import ApiKeys from "./admin/settings/ApiKeys.vue";
+import SetupWizard from "./admin/setup/SetupWizard.vue";
+import Plugins from "./admin/plugins/Plugins.vue";
 import "./admin/admin.scss";
 import "./ui/theme.css";
 import { initTheme, loadThemes } from "./ui/theme";
 import { ws } from "./api/ws";
-import { initLogStream } from "./admin/state/logStream";
+import { initLogStream } from "./admin/logs/logStream";
 import { setup as setupApi } from "./api/client";
 
 async function main() {
